Reorganized files to accomodate mobile version of table. Also @ts-ignored a bunch of stuff :D
